requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.18.0
protobuf==3.11.2
scikit-learn==0.22
scipy==1.3.1
six==1.12.0
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.16.0
wincertstore==0.2
WTForms==2.2.1
opencv-python==4.4.0
gevent==1.2.2
pillow>=6.2.0